pyproject.toml
[project]
name = "layout-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.116.1",
    "paddleocr==2.6.1.3",
    "paddlepaddle==2.5.0",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "setuptools>=80.9.0",
    "uvicorn[standard]>=0.35.0",
]
